Move basic object tests to their own file
